requirements.txt
fastapi
uvicorn
jinja2
python-multipart
opencv-python
numpy
Pillow
